Import - Add UserJobTrait
