+++ CRM/Import/Forms.php
@@ -25,22 +25,12 @@ use League\Csv\Writer;
  * This class helps the forms within the import flow access submitted & parsed values.
  */
 class CRM_Import_Forms extends CRM_Core_Form {
-
+  use \Civi\UserJob\UserJobTrait;
 
   /**
    * @var int
    */
   protected $templateID;
-
-  /**
-   * User job id.
-   *
-   * This is the primary key of the civicrm_user_job table which is used to
-   * track the import.
-   *
-   * @var int
-   */
-  protected $userJobID;
 
   /**
    * Name of the import mapping (civicrm_mapping).
@@ -87,36 +77,6 @@ class CRM_Import_Forms extends CRM_Core_Form {
   }
 
   /**
-   * @return int|null
-   */
-  public function getUserJobID(): ?int {
-    if (!$this->userJobID && $this->get('user_job_id')) {
-      $this->userJobID = $this->get('user_job_id');
-    }
-    return $this->userJobID;
-  }
-
-  /**
-   * Set user job ID.
-   *
-   * @param int $userJobID
-   */
-  public function setUserJobID(int $userJobID): void {
-    $this->userJobID = $userJobID;
-    // This set allows other forms in the flow ot use $this->get('user_job_id').
-    $this->set('user_job_id', $userJobID);
-  }
-
-  /**
-   * User job details.
-   *
-   * This is the relevant row from civicrm_user_job.
-   *
-   * @var array
-   */
-  protected $userJob;
-
-  /**
    * @var \CRM_Import_Parser
    */
   protected $parser;
@@ -130,25 +90,6 @@ class CRM_Import_Forms extends CRM_Core_Form {
    * @var bool
    */
   protected $isQuickFormMode = TRUE;
-
-  /**
-   * Get User Job.
-   *
-   * API call to retrieve the userJob row.
-   *
-   * @return array
-   *
-   * @throws \CRM_Core_Exception
-   */
-  protected function getUserJob(): array {
-    if (!$this->userJob) {
-      $this->userJob = UserJob::get()
-        ->addWhere('id', '=', $this->getUserJobID())
-        ->execute()
-        ->first();
-    }
-    return $this->userJob;
-  }
 
   /**
    * Get submitted values stored in the user job.
